feat: move server settings into feature and page
